Separated Database stuff into a library and linked the library to the WPF project
This way we can refine the library and then create a ui overlay with any
kind of UI Library. If need be we can create a WinForms project that is
linked to the BlurayBackupManagement library.If the library is updated
then we can just rebuild the solution and any kind of project can be
built with the updated functions of the library directly.